Added NXT sensor-type Active Light Sensor Deactivate all sensors on stage-destroy and Mindstorm-disconnect Fixed duration of PlayTone

+++ catroid/src/org/catrobat/catroid/devices/mindstorms/nxt/LegoNXTImpl.java
@@ -86,7 +86,11 @@ public class LegoNXTImpl implements LegoNXT, NXTSensorService.OnSensorChangedLis
 	public void disconnect() {
 		if (mindstormsConnection.isConnected()) {
 			this.stopAllMovements();
-			sensorService.destroy();
+
+			if (sensorService != null) {
+				sensorService.deactivateAllSensors(mindstormsConnection);
+				sensorService.destroy();
+			}
 			mindstormsConnection.disconnect();
 		}
 	}
@@ -117,9 +121,9 @@ public class LegoNXTImpl implements LegoNXT, NXTSensorService.OnSensorChangedLis
 	}
 
 	@Override
-	public void playTone(int frequencyInHz, int durationInMs) {
+	public void playTone(int frequencyInHz, int durationInS) {
 
-		if (durationInMs <= 0) {
+		if (durationInS <= 0) {
 			return;
 		}
 
@@ -129,6 +133,8 @@ public class LegoNXTImpl implements LegoNXT, NXTSensorService.OnSensorChangedLis
 		else if (frequencyInHz < 200) {
 			frequencyInHz = 200;
 		}
+
+		int durationInMs = durationInS * 1000;
 
 		Command command = new Command(CommandType.DIRECT_COMMAND, CommandByte.PLAY_TONE, false);
 		command.append((byte)(frequencyInHz & 0x00FF));
@@ -242,7 +248,7 @@ public class LegoNXTImpl implements LegoNXT, NXTSensorService.OnSensorChangedLis
 		motorB = new NXTMotor(1, mindstormsConnection);
 		motorC = new NXTMotor(2, mindstormsConnection);
 
-		assignSensorsToPorts();
+
 
 		isInitialized = true;
 	}
@@ -259,16 +265,21 @@ public class LegoNXTImpl implements LegoNXT, NXTSensorService.OnSensorChangedLis
 	@Override
 	public void start() {
 		initialise();
+
+		assignSensorsToPorts();
+
 		sensorService.resumeSensorUpdate();
 	}
 
 	@Override
-	public void pause() {
+	public void pause()
+	{
 		stopAllMovements();
 		sensorService.pauseSensorUpdate();
 	}
 
 	@Override
 	public void destroy() {
+		sensorService.deactivateAllSensors(mindstormsConnection);
 	}
 }
